Relocating diagnostic primitive.
