+++ src/diagnostics/primitives/consoleWrapper.ts
